refactor: move enrol logic to schedule component
