+++ lib/safira_web/live/landing/components/schedule.ex
@@ -2,22 +2,23 @@ defmodule SafiraWeb.Landing.Components.Schedule do
   @moduledoc """
   Schedule component.
   """
-  use SafiraWeb, :component
+  use SafiraWeb, :live_component
 
   alias Plug.Conn.Query
   alias Safira.Activities
 
-  attr :event_start_date, Date, required: true
-  attr :event_end_date, Date, required: true
-  attr :url, :string, required: true
-  attr :params, :map, required: true
-  attr :has_filters?, :boolean, default: false
-  attr :descriptions_enabled, :boolean, default: false
+  @impl true
+  def mount(socket) do
+    {:ok, socket}
+  end
 
-  attr :user_role, :atom, required: true
-  attr :enrolments, :list, required: true
+  @impl true
+  def update(assigns, socket) do
+    {:ok, socket |> assign(assigns) |> assign(enrolments: get_enrolments(assigns.current_user))}
+  end
 
-  def schedule(assigns) do
+  @impl true
+  def render(assigns) do
     ~H"""
     <div class="xl:grid 2xl:grid-cols-2 gap-8 relative select-none">
       <div class="mb-20 2xl:mb-0">
@@ -43,8 +44,9 @@ defmodule SafiraWeb.Landing.Components.Schedule do
           date={fetch_current_date_from_params(assigns.params) || assigns.event_start_date}
           filters={fetch_filters_from_params(assigns.params)}
           descriptions_enabled={assigns.descriptions_enabled}
-          user_role={assigns.user_role}
+          user_role={get_user_role(assigns.current_user)}
           enrolments={assigns.enrolments}
+          myself={assigns.myself}
         />
       </div>
     </div>
@@ -94,6 +96,7 @@ defmodule SafiraWeb.Landing.Components.Schedule do
                 descriptions_enabled={@descriptions_enabled}
                 user_role={@user_role}
                 enrolments={@enrolments}
+                myself={@myself}
               />
             <% end %>
           <% end %>
@@ -168,6 +171,7 @@ defmodule SafiraWeb.Landing.Components.Schedule do
                 :if={enrolments_enabled(@activity, @user_role, @enrolments)}
                 class="relative hover:underline cursor-pointer -mr-3 font-iregular text-lg text-accent sm:mr-1"
                 phx-click="enrol"
+                phx-target={@myself}
                 data-confirm={"#{gettext("You are enrolling for")} #{@activity.title}. #{gettext("This action cannot be undone. Are you sure?")}"}
                 phx-value-activity_id={@activity.id}
               >
@@ -349,6 +353,58 @@ defmodule SafiraWeb.Landing.Components.Schedule do
       </defs>
     </svg>
     """
+  end
+
+  @impl true
+  def handle_event("enrol", %{"activity_id" => activity_id}, socket) do
+    if is_nil(socket.assigns.current_user) do
+      {:noreply,
+       socket
+       |> put_flash(:error, gettext("You must be logged in to enrol in activities"))
+       |> redirect(to: ~p"/users/log_in?action=enrol&action_id=#{activity_id}&return_to=/")}
+    else
+      if socket.assigns.current_user.type == :attendee do
+        actual_enrol(activity_id, socket)
+      else
+        {:noreply,
+         socket
+         |> put_flash(:error, gettext("Only attendees can enrol in activities"))}
+      end
+    end
+  end
+
+  defp actual_enrol(activity_id, socket) do
+    case Activities.enrol(socket.assigns.current_user.attendee.id, activity_id) do
+      {:ok, _} ->
+        send(self(), {:update_flash, {:info, gettext("Successfully enrolled")}})
+
+        {:noreply,
+         socket
+         |> assign(
+           :enrolments,
+           Activities.get_attendee_enrolments(socket.assigns.current_user.attendee.id)
+         )}
+
+      {:error, _} ->
+        send(self(), {:update_flash, {:info, gettext("Unable to enrol")}})
+        {:noreply, socket}
+    end
+  end
+
+  defp get_enrolments(user) do
+    if is_nil(user) or user.type != :attendee do
+      []
+    else
+      Activities.get_attendee_enrolments(user.attendee.id)
+    end
+  end
+
+  defp get_user_role(user) do
+    if is_nil(user) do
+      :attendee
+    else
+      user.type
+    end
   end
 
   defp fetch_current_date_from_params(params) do
