+++ lib/safira_web/live/landing/home_live/index.ex
@@ -3,7 +3,6 @@ defmodule SafiraWeb.Landing.HomeLive.Index do
   use SafiraWeb, :landing_view
 
   import SafiraWeb.Landing.HomeLive.Components.{Hero, Partners, Pitch, Sponsors, Speakers}
-  import SafiraWeb.Landing.Components.Schedule
 
   alias Safira.{Activities, Event}
 
@@ -20,8 +19,6 @@ defmodule SafiraWeb.Landing.HomeLive.Index do
      |> assign(:registrations_open?, Event.registrations_open?())
      |> assign(:has_sponsors?, Companies.get_companies_count() > 0)
      |> assign(:has_schedule?, Activities.get_activities_count() > 0)
-     |> assign(:enrolments, get_enrolments(socket.assigns.current_user))
-     |> assign(:user_role, get_user_role(socket.assigns.current_user))
      |> stream(:speakers, speakers |> Enum.shuffle())}
   end
 
@@ -31,54 +28,7 @@ defmodule SafiraWeb.Landing.HomeLive.Index do
   end
 
   @impl true
-  def handle_event("enrol", %{"activity_id" => activity_id}, socket) do
-    if is_nil(socket.assigns.current_user) do
-      {:noreply,
-       socket
-       |> put_flash(:error, gettext("You must be logged in to enrol in activities"))
-       |> redirect(to: ~p"/users/log_in?action=enrol&action_id=#{activity_id}&return_to=/")}
-    else
-      if socket.assigns.current_user.type == :attendee do
-        actual_enrol(activity_id, socket)
-      else
-        {:noreply,
-         socket
-         |> put_flash(:error, gettext("Only attendees can enrol in activities"))}
-      end
-    end
-  end
-
-  defp get_enrolments(user) do
-    if is_nil(user) or user.type != :attendee do
-      []
-    else
-      Activities.get_attendee_enrolments(user.attendee.id)
-    end
-  end
-
-  defp actual_enrol(activity_id, socket) do
-    case Activities.enrol(socket.assigns.current_user.attendee.id, activity_id) do
-      {:ok, _} ->
-        {:noreply,
-         socket
-         |> put_flash(:info, gettext("Successfully enrolled"))
-         |> assign(
-           :enrolments,
-           Activities.get_attendee_enrolments(socket.assigns.current_user.attendee.id)
-         )}
-
-      {:error, _} ->
-        {:noreply,
-         socket
-         |> put_flash(:error, gettext("Something happened"))}
-    end
-  end
-
-  defp get_user_role(user) do
-    if is_nil(user) do
-      :attendee
-    else
-      user.type
-    end
+  def handle_info({:update_flash, {flash_type, msg}}, socket) do
+    {:noreply, put_flash(socket, flash_type, msg)}
   end
 end
