+++ lib/safira_web/live/landing/home_live/index.html.heex
@@ -7,14 +7,17 @@
   <.pitch :if={@has_schedule? || @has_highlighted_speakers? || @has_sponsors?} />
 </div>
 <div class="px-5 md:px-32 xl:px-16 xl:px-32 pt-20">
-  <.schedule
+  <.live_component
     :if={@has_schedule?}
+    id={:schedule}
+    module={SafiraWeb.Landing.Components.Schedule}
     event_start_date={@event_start_date}
     event_end_date={@event_end_date}
     url={~p"/"}
     params={@params}
-    enrolments={@enrolments}
-    user_role={@user_role}
+    has_filters?={false}
+    descriptions_enabled={false}
+    current_user={@current_user}
   />
 </div>
 <div class="xl:px-[15rem] md:px-[8rem] px-[2.5rem]">
