+++ lib/safira_web/live/landing/schedule_live/index.html.heex
@@ -13,15 +13,16 @@
     </h1>
   </div>
   <div class="mt-40">
-    <.schedule
+    <.live_component
+      id={:schedule}
+      module={SafiraWeb.Landing.Components.Schedule}
       event_start_date={@event_start_date}
       event_end_date={@event_end_date}
-      url={~p"/schedule"}
-      params={@params}
+      url={~p"/"}
       has_filters?={true}
       descriptions_enabled={true}
-      enrolments={@enrolments}
-      user_role={@user_role}
+      params={@params}
+      current_user={@current_user}
     />
   </div>
 </div>
